Revamping the files, must make the tests.
