+++ context_managers/working.py
@@ -16,7 +16,7 @@ from typing import Any, Union
 
 
 # /////////////////////////////////////////////////////////////////////////////
-# Working Directory
+# Classes
 # /////////////////////////////////////////////////////////////////////////////
